hide xyzPRUConfig() models
