+++ shepherd_core/shepherd_core/vsource/virtual_source_model.py
@@ -12,10 +12,10 @@ NOTE: DO NOT OPTIMIZE -> stay close to original code-base
 
 from shepherd_core.data_models.base.calibration import CalibrationEmulator
 from shepherd_core.data_models.content.enum_datatypes import EnergyDType
-from shepherd_core.data_models.content.virtual_harvester_config import HarvesterPRUConfig
-from shepherd_core.data_models.content.virtual_source_config import ConverterPRUConfig
+from shepherd_core.data_models.content.virtual_harvester_config_pru import HarvesterPRUConfig
 from shepherd_core.data_models.content.virtual_source_config import VirtualSourceConfig
-from shepherd_core.data_models.content.virtual_storage_config import StoragePRUConfig
+from shepherd_core.data_models.content.virtual_source_config_pru import ConverterPRUConfig
+from shepherd_core.data_models.content.virtual_storage_config_pru import StoragePRUConfig
 
 from .virtual_converter_model import PruCalibration
 from .virtual_converter_model import VirtualConverterModel
