+++ shepherd_core/shepherd_core/vsource/virtual_storage_model_fixed_point_math.py
@@ -3,7 +3,7 @@
 from itertools import product
 
 from shepherd_core import log
-from shepherd_core.data_models.content.virtual_storage_config import LuT_SIZE_LOG
+from shepherd_core.data_models.content.virtual_storage_config_pru import LuT_SIZE_LOG
 
 
 def u32l(i: float) -> int:
