+++ shepherd_core/tests/vsource/test_harvester.py
@@ -3,7 +3,7 @@ from pathlib import Path
 import pytest
 from shepherd_core.data_models import EnergyDType
 from shepherd_core.data_models import VirtualHarvesterConfig
-from shepherd_core.data_models.content.virtual_harvester_config import HarvesterPRUConfig
+from shepherd_core.data_models.content.virtual_harvester_config_pru import HarvesterPRUConfig
 from shepherd_core.vsource import VirtualHarvesterModel
 
 from shepherd_core import Reader
